Rename TheButton, TheCheckBox to BaseButton, BaseCheckBox
