+++ project/frontend/src/components/UI/BaseButton.vue
@@ -1,10 +1,20 @@
 <template>
-    <button>{{ text }}</button>
+    <button :class="mode">{{ text }}</button>
 </template>
 
 <script>
 export default {
-    props: ['text'],
+    props: {
+        mode: {
+            type: String,
+            required: false,
+            default: 'highlight',
+        },
+        text: {
+            type: String,
+            required: false,
+        }
+    },
 }
 </script>
 
@@ -17,14 +27,21 @@ button {
     color: #000;
     border-radius: 8px;
     border: 1px solid #ffffff;
-    box-shadow: 0 1px 4px rgba(0, 0, 0, 0.26);
     background-color: #ffffff;
     padding: 1rem 3rem;
     outline: none;
     cursor: pointer;
 }
 
-button:focus {
+.flat {
+    
+}
+
+.highlight {
+    box-shadow: 0 1px 4px rgba(0, 0, 0, 0.26);
+}
+
+.highlight:focus {
     box-shadow: 0 2px 8px rgba(0, 0, 0, 0.26);
 }
 </style>
