+++ project/frontend/src/main.js
@@ -3,14 +3,12 @@ import App from './App.vue'
 import router from './router'
 import store from './store'
 
-import TheHeader from '@/components/layout/TheHeader.vue';
-import TheButton from '@/components/layout/TheButton.vue';
-import TheCheckBox from '@/components/layout/TheCheckBox.vue';
+import BaseButton from '@/components/UI/BaseButton.vue';
+import BaseCheckBox from '@/components/UI/BaseCheckBox.vue';
 
 const app = createApp(App);
-app.component('TheHeader', TheHeader);
-app.component('TheButton', TheButton);
-app.component('TheCheckBox', TheCheckBox);
+app.component('BaseButton', BaseButton);
+app.component('BaseCheckBox', BaseCheckBox);
 app.use(store);
 app.use(router);
 app.mount('#app');
